asciidoc/courses/workshop-microsoft-agents-framework/lab-repository/requirements.txt
agent-framework-core==1.0.0rc2
agent-framework-neo4j>=0.5.4
neo4j-agent-memory @ git+https://github.com/neo4j-labs/agent-memory.git@workshop-temperature
neo4j>=5.20.0
neo4j-graphrag>=1.13.0
openai>=1.99.0
opentelemetry-semantic-conventions-ai==0.4.13
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
spacy>=3.0.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
jupyter>=1.0.0
ipykernel>=6.0.0
